rename staedystate file and exact method
